+++ datumaro/cli/contexts/project/__init__.py
@@ -18,7 +18,6 @@ from datumaro.components.operations import (DistanceComparator,
 from datumaro.components.project import \
     PROJECT_DEFAULT_CONFIG as DEFAULT_CONFIG
 from datumaro.components.project import Environment, Project
-from datumaro.components.validator import validate_annotations, TaskType
 from datumaro.util import error_rollback
 
 from ...util import (CliException, MultilineFormatter, add_subparser,
@@ -793,64 +792,6 @@ def info_command(args):
 
     return 0
 
-def build_validate_parser(parser_ctor=argparse.ArgumentParser):
-    parser = parser_ctor(help="Validate project",
-        description="""
-            Validates project based on specified task type and stores
-            results like statistics, reports and summary in JSON file.
-        """,
-        formatter_class=MultilineFormatter)
-
-    parser.add_argument('task_type',
-        choices=[task_type.name for task_type in TaskType],
-        help="Task type for validation")
-    parser.add_argument('-s', '--subset', dest='subset_name', default=None,
-        help="Subset to validate (default: None)")
-    parser.add_argument('-p', '--project', dest='project_dir', default='.',
-        help="Directory of the project to validate (default: current dir)")
-    parser.add_argument('extra_args', nargs=argparse.REMAINDER, default=None,
-        help="Optional arguments for validator (pass '-- -h' for help)")
-    parser.set_defaults(command=validate_command)
-
-    return parser
-
-def validate_command(args):
-    project = load_project(args.project_dir)
-    task_type = args.task_type
-    subset_name = args.subset_name
-    dst_file_name = f'validation_results-{task_type}'
-
-    dataset = project.make_dataset()
-    if subset_name is not None:
-        dataset = dataset.get_subset(subset_name)
-        dst_file_name += f'-{subset_name}'
-
-    extra_args = {}
-    from datumaro.components.validator import _Validator
-    extra_args = _Validator.parse_cmdline(args.extra_args)
-    validation_results = validate_annotations(dataset, task_type, **extra_args)
-
-    def numpy_encoder(obj):
-        if isinstance(obj, np.generic):
-            return obj.item()
-
-    def _make_serializable(d):
-        for key, val in list(d.items()):
-            # tuple key to str
-            if isinstance(key, tuple):
-                d[str(key)] = val
-                d.pop(key)
-            if isinstance(val, dict):
-                _make_serializable(val)
-
-    _make_serializable(validation_results)
-
-    dst_file = generate_next_file_name(dst_file_name, ext='.json')
-    log.info("Writing project validation results to '%s'" % dst_file)
-    with open(dst_file, 'w') as f:
-        json.dump(validation_results, f, indent=4, sort_keys=True,
-                  default=numpy_encoder)
-
 def build_parser(parser_ctor=argparse.ArgumentParser):
     parser = parser_ctor(
         description="""
@@ -873,6 +814,5 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
     add_subparser(subparsers, 'transform', build_transform_parser)
     add_subparser(subparsers, 'info', build_info_parser)
     add_subparser(subparsers, 'stats', build_stats_parser)
-    add_subparser(subparsers, 'validate', build_validate_parser)
 
     return parser
